+++ VK Client/Authorization/LoginViewController.swift
@@ -58,7 +58,6 @@ class LoginViewController: UIViewController {
         return checkResult
     }
     
-// Переименовал:
     func isUserDataValid() -> Bool {
         
         let login = loginTextField.text
@@ -91,37 +90,5 @@ class LoginViewController: UIViewController {
     
 // Метод возврата на экран авторизации:
     @IBAction func myUnwindAction(segue: UIStoryboardSegue) {
-    }
-}
-
-extension LoginViewController {
-    
-    @objc func textFieldDidChange(textField: UITextField) {
-        if loginTextField.text!.isEmpty || passwordTextField.text!.isEmpty {
-            signInButton.isEnabled = false
-        } else {
-            signInButton.isEnabled = true
-        }
-    }
-    
-    @objc func keyboardWasShown(notification: Notification) {
-        
-        let info = notification.userInfo! as NSDictionary
-        let kbSize = (info.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as! NSValue).cgRectValue.size
-        let contentInsets = UIEdgeInsets(top: 0, left: 0, bottom: kbSize.height, right: 0)
-        
-        self.scrollView.contentInset = contentInsets
-        scrollView.scrollIndicatorInsets = contentInsets
-    }
-    
-    @objc func keyboardWillBeHidden(notification: Notification) {
-        
-        let contentInsets = UIEdgeInsets.zero
-        scrollView.contentInset = contentInsets
-    }
-
-    
-    @objc func hideKeyboard() {
-        self.scrollView.endEditing(true)
     }
 }
