+++ VK Client/Groups/MyGroups/MyGroupsCell.swift
@@ -21,6 +21,7 @@ class MyGroupsCell: UITableViewCell {
     }
     
     func configure(for model: Group) {
+        
         myGroupNameLabel.text = model.nameGroup
         myGroupImage.image = UIImage(named: model.imageGroup)
     }
