+++ VK Client/Supporting Files/AppDelegate.swift
