+++ VK Client/Supporting Files/Info.plist
