+++ VK Client/Supporting Files/SceneDelegate.swift
